% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{contrib_approx}
\alias{contrib_approx}
\title{contrib_approx}
\usage{
contrib_approx(emissions, type = "country", entity)
}
\arguments{
\item{emissions}{country or group level emissions dataset}

\item{type}{country or group}

\item{entity}{iso3 country or unfccc group code}
}
\value{
a one-row data-frame giving gsat_loo, gsat_loi and gsat_approx
}
\description{
Approximation to country or group level contribution to global warming based on LOO and LOI values for the entity.
}
\details{
Set type=country if working with a country-level emissions dataset. Set type=group if working with a grouped emissions
dataset created by allocateR::regroup_emissions. This allows comparison with output produced by allocateR::contrib_shap which
requires grouped data
}
