% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{contrib_shap_p}
\alias{contrib_shap_p}
\title{contrib_shap_p}
\usage{
contrib_shap_p(emissions, group)
}
\arguments{
\item{emissions}{a grouped emissions dataset such as that produced by allocateR::regroup_emissions()}

\item{group}{a country (iso3) or unfccc group code}
}
\value{
a reasonable gsat allocation value.
}
\description{
evaluate the reasonable allocation GSAT of country or grouping given an emissions dataset and a list of all possible coalitions.
contrib_shap_p is the parallel version of contrib_shap using mclapply
}
