% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_cumulative_ffi_emitters.R
\docType{data}
\name{country_cumulative_ffi_shares}
\alias{country_cumulative_ffi_shares}
\title{Ranked Cumulative CO2 Emitters}
\format{
A data frame with 195 rows and 3 columns:
\describe{
  \item{country}{country}
  \item{cumulative}{Cumulative Pg C emissions}
  \item{share}{percentage share of cumulative ffi emissions}
  }
}
\source{
CEDS
}
\usage{
country_cumulative_ffi_shares
}
\description{
Based on CEDS 2022 ffi emissions. Include eu member states under "eu27".
}
\keyword{datasets}
