% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_ceds.R
\docType{data}
\name{emissions_ceds}
\alias{emissions_ceds}
\title{CEDS emissions data at country level}
\format{
## unfccc_groupings
A data frame with 38120 rows and 5 columns:
\describe{
  \item{country}{iso code}
  \item{year}{year}
  \item{variable}{climate forcer}
  \item{value}{emissions amount}
  \item{units}{units}
}
}
\source{
<https://zenodo.org/records/10904361>
}
\usage{
emissions_ceds
}
\description{
Annual 1851-2022 emissions of FFI-CO2, CH4, N2O, BC, OC, SO2, CO, NOX, NMVOC, NH3
LULUCF emissions are not included.
}
\details{
CEDS data for CH4 and N2O do not extend for 1970 so these are generated using global estimates.
}
\keyword{datasets}
