% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{get_coalition_emissions}
\alias{get_coalition_emissions}
\title{get_coalition_emissions}
\usage{
get_coalition_emissions(emissions)
}
\arguments{
\item{emissions}{emissions of individual countrie/groups in a coalition}
}
\value{
emissions of the coalition
}
\description{
helper function to combine emissions in a coalition
}
