% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{get_coalitions}
\alias{get_coalitions}
\title{get_coalitions}
\usage{
get_coalitions(groups)
}
\arguments{
\item{groups}{a vector of countries or negotiating groups}
}
\value{
a list of length 2^N
}
\description{
Helper function that creates a list of all coalitions from a list of countries or country groupings. If there are N groups the number of coalitions is 2^N including
the null coalition.
}
