% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_cumulative_ffi_shares.R
\docType{data}
\name{group_cumulative_ffi_shares}
\alias{group_cumulative_ffi_shares}
\title{group_cumulative_ffi_shares}
\format{

}
\source{
<https://zenodo.org/records/10904361>
}
\usage{
group_cumulative_ffi_shares
}
\description{
Cumulative FFI CO2 emissions by unfccc negotiating group using assignment of cohesion model
}
\keyword{datasets}
