% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{group_emissions}
\alias{group_emissions}
\title{group_emissions}
\usage{
group_emissions(emissions, groupings)
}
\arguments{
\item{emissions}{historical emissions at country level in Hector format}

\item{groupings}{groupings dataframe coontaining country and grouping codes}
}
\value{
emissions dataframe at grouping level in Hector format
}
\description{
Takes a country level historical emissions dataset and converts to a grouped emissions dataset
}
\examples{
"group_emissions(emissions_ceds,unfccc_groupings_cohesion)"
}
