% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{regroup_emissions}
\alias{regroup_emissions}
\title{regroup_emissions}
\usage{
regroup_emissions(
  emissions,
  add_countries,
  drop_groupings = c("cvf", "mp", "g77", "eig", "cacam", "sids", "ailac", "alba", "rn")
)
}
\arguments{
\item{emissions}{country-level emissions dataset with original grouping labels (based on e.g. unfccc_groupings_cohesion)}

\item{add_countries}{Countries to separate out that their gsat contribution can be calculated individually}

\item{drop_groupings}{Groups to assign to the "other" category}
}
\value{
a regrouped emissions dataset with grouping code and no country codes
}
\description{
regroup_emissions regroups a country-level emissions dataset, allowing the impact of selected individual countries to be found. Groups in drop_groupings are moved into the "other" grouping category.
A list of countries add_countries are separated into their own groups with grouping label equal to the country code. The returned dataframe
has emissions aggregated by the new grouping labels and can be used directly with contrib_shap.
}
