% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{rollup_emissions}
\alias{rollup_emissions}
\title{rollup_emissions}
\usage{
rollup_emissions(countrynames, group_eu = TRUE)
}
\arguments{
\item{countrynames}{a vector of iso3 country codes (may include eu27 when group_eu is TRUE)}

\item{group_eu}{If TRUE groups EU member states to "eu27"}
}
\value{
an emissions data-frame for countrynames and "rem"
}
\description{
re-expresses CEDS country level emissions data as the emissions of countries in countrynames plus a remainder.
option to group eu27 member states if desired.
}
