% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfccc_groupings.R
\docType{data}
\name{unfccc_groups}
\alias{unfccc_groups}
\title{UNFCCC Negotiating Group Names}
\format{
## unfccc_groupings
A data frame with 20 rows and 2 columns:
\describe{
  \item{grouping}{grouping code}
  \item{name}{Name of negotiating group}
  #' }
}
\source{
<https://unfccc.int/process-and-meetings/parties-non-party-stakeholders/parties/party-groupings>
}
\usage{
unfccc_groups
}
\description{
The names and corresponding code
}
\keyword{datasets}
