% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{contrib_shap}
\alias{contrib_shap}
\title{contrib_shap}
\usage{
contrib_shap(emissions, country, coalitions)
}
\arguments{
\item{emissions}{the emissions dataset for all countries or groupings}

\item{country}{the country (iso3) or grouping code}

\item{coalitions}{the list of all coalitions of length 2^N where N is the number of countries or groupings}
}
\value{
the reasonable gsat allocation for the country
}
\description{
evaluate the reasonable allocation GSAT of country or grouping given an emissions dataset and a list of all possible coalitions
}
