% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{make_null_core}
\alias{make_null_core}
\title{make_null_core}
\usage{
make_null_core(natural_n2o = 9.96, natural_ch4 = 337.7725)
}
\arguments{
\item{natural_n2o}{Tuned natural N2O emissions}

\item{natural_ch4}{Tuned natural CH4 emissions}
}
\value{
a hector core
}
\description{
A hector model instance (core) with zeroed anthropogenic and volcanic sulfur emissions
}
