% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{regroup_emissions}
\alias{regroup_emissions}
\title{regroup_emissions}
\usage{
regroup_emissions(
  emissions,
  countries_add = c("nzl", "irl", "ury"),
  groupings_sub = c("cvf", "mp", "g77", "eig", "cacam", "sids", "ailac", "alba", "rn")
)
}
\arguments{
\item{emissions}{country-level emissions dataset with original grouping labels (based on e.g. unfccc_groupings_cohesion)}

\item{countries_add}{A vector of countries to add so that their gsat contribution can be calculated.}

\item{groupings_sub}{A vector of groups to subtract i.e. to assign to the "other" category}
}
\value{
a regrouped emissions dataset with grouping code and no country codes
}
\description{
regroup_emissions regroups a country-level emissions dataset, allowing the warming impact of selected individual countries to be found. Groups in groupings_drop are moved into the "other" grouping category.
Countries in a vector of iso codes countries_add appear with group label equal to the country code. The returned dataframe
has emissions aggregated by the new grouping labels and can be used with using allocateR::contrib_shap.
}
\details{
IMPORTANT LIMITATION. regroup_emissions is not intended to work when more than two or more of the selected countries belong to the same unfccc grouping.
}
